Make test data

Make a toy data set for testing and demo. This is for internal use purpose and not intended to be called by users.

Usage

make_test_dat(
  vals_kept = c("304", "305", 3040:3049, 3050:3059),
  noise_val = "999",
  IDs = 1:50,
  date_range = seq(as.Date("2015-01-01"), as.Date("2020-12-31"), by = 1),
  nrows = 100,
  n_any = 50,
  n_all = 10,
  seed = NULL,
  answer_id = NULL,
  type = c("data.frame", "database")
)

Arguments

vals_kept

A vector of values that suppose to be identified.

noise_val

A vector of values that are not meant to be identified.

IDs

A vector of client IDs.

date_range

A vector of all possible dates in the data.

nrows

Number of rows of the output.

n_any

Number of rows to be identified if the criteria is that if any target column contains certain values.

n_all

Number of rows to be identified if the criteria is that if all target columns contain certain values.

seed

Seed for random number generation.

answer_id

Column name for the indicator of how the row should be identified: any, all, and noise.

type

Output type, "data.frame" or "database".

Value

A data.frame or remote table from 'dbplyr'.

Examples

make_test_dat()
#>     uid clnt_id      dates diagx diagx_1 diagx_2
#> 1    56       1 2016-06-21   999    <NA>     999
#> 2    77       1 2016-08-18   999    <NA>     999
#> 3    93       1 2018-11-17   999    <NA>    <NA>
#> 4    91       1 2019-01-23   999     999     999
#> 5     5       3 2015-09-16   304    3057     999
#> 6    79       3 2019-02-14   999    <NA>    <NA>
#> 7    80       3 2020-01-27   999    <NA>     999
#> 8    16       3 2020-02-19  3043    3053     999
#> 9     3       4 2015-11-01  3057    3057    <NA>
#> 10   58       4 2019-05-22   999     999     999
#> 11   62       5 2015-08-21   999    <NA>    <NA>
#> 12   47       5 2019-08-24  3041    3054    3047
#> 13   18       7 2017-06-19  3059    3051     999
#> 14   57       7 2019-09-19   999     999     999
#> 15   71       7 2020-07-05   999     999     999
#> 16   81       8 2016-03-06   999    <NA>    <NA>
#> 17   30       8 2019-02-05  3051     999     999
#> 18   83       9 2015-07-16   999    <NA>     999
#> 19   54       9 2015-12-19   999     999     999
#> 20   42       9 2018-02-20  3048    3047    3043
#> 21   37       9 2018-11-10  3046    3048     999
#> 22   13       9 2019-07-09  3044    3051    <NA>
#> 23   55      10 2015-08-06   999     999     999
#> 24   76      10 2016-05-07   999     999     999
#> 25   20      10 2018-03-07   305    3040    <NA>
#> 26   66      10 2019-06-25   999    <NA>    <NA>
#> 27   25      10 2020-02-24  3049    3045    <NA>
#> 28   44      11 2016-04-29   305    3048    3042
#> 29   36      12 2019-07-26  3048    3049     999
#> 30   85      13 2019-01-21   999     999    <NA>
#> 31   64      13 2020-09-05   999     999    <NA>
#> 32   24      16 2015-07-17  3042    3042    <NA>
#> 33  100      16 2016-07-30   999     999    <NA>
#> 34   49      16 2018-12-20  3047    3055     305
#> 35   87      18 2015-11-23   999     999     999
#> 36   75      18 2017-02-12   999     999    <NA>
#> 37   11      20 2018-05-20  3040    3056    <NA>
#> 38   69      20 2019-08-27   999     999    <NA>
#> 39   73      20 2020-02-07   999     999    <NA>
#> 40   51      21 2020-04-28   999    <NA>    <NA>
#> 41   32      22 2016-10-09   305    3059     999
#> 42   27      22 2019-05-16  3058    3052    <NA>
#> 43   84      22 2020-05-14   999    <NA>    <NA>
#> 44   67      23 2015-08-15   999     999     999
#> 45    4      23 2016-01-12  3044    3059     999
#> 46   26      23 2018-07-26  3040    3040     999
#> 47   48      23 2019-10-17  3043    3049    3043
#> 48   70      24 2016-01-01   999     999    <NA>
#> 49   92      24 2016-07-28   999     999    <NA>
#> 50   29      24 2017-04-01  3048    3042     999
#> 51   68      24 2017-08-19   999    <NA>    <NA>
#> 52   52      24 2017-10-15   999     999     999
#> 53   95      25 2018-01-26   999    <NA>     999
#> 54   43      26 2020-01-22  3047    3052     304
#> 55   33      27 2018-10-22  3057    3051    <NA>
#> 56   90      27 2019-01-08   999     999     999
#> 57   45      27 2019-10-10  3054     304    3051
#> 58    8      28 2016-05-27  3048    3043    <NA>
#> 59   94      28 2020-12-13   999    <NA>    <NA>
#> 60   72      29 2015-06-12   999     999     999
#> 61   39      29 2016-06-06  3041    3043    <NA>
#> 62    2      31 2017-07-12  3051     305    <NA>
#> 63   17      31 2018-01-30  3051    3040     999
#> 64   96      33 2015-09-24   999     999     999
#> 65   99      34 2016-07-20   999    <NA>    <NA>
#> 66   65      34 2020-10-19   999     999    <NA>
#> 67    9      35 2016-01-13  3056    3046    <NA>
#> 68    6      35 2016-05-02   304    3046     999
#> 69   78      35 2017-04-20   999    <NA>     999
#> 70   46      35 2018-11-18  3042    3058    3041
#> 71   63      36 2016-04-23   999    <NA>    <NA>
#> 72   88      36 2018-01-04   999    <NA>     999
#> 73   21      37 2016-12-18   305    3048    <NA>
#> 74   59      37 2017-03-19   999    <NA>    <NA>
#> 75   53      38 2015-03-15   999    <NA>     999
#> 76   60      38 2018-02-26   999    <NA>     999
#> 77   41      39 2017-02-18   305    3054    3049
#> 78   40      40 2015-08-15  3053    3055     999
#> 79   14      40 2018-03-03  3053    3050    <NA>
#> 80   10      40 2019-07-12  3046    3055     999
#> 81   28      41 2016-02-25  3051    3044     999
#> 82   31      41 2019-10-03  3048    3055    <NA>
#> 83   12      42 2018-03-19  3044    3052     999
#> 84   89      43 2020-08-28   999     999    <NA>
#> 85   74      44 2016-04-09   999    <NA>    <NA>
#> 86   35      45 2018-06-21  3051    3058    <NA>
#> 87    7      45 2020-07-12  3045    3042    <NA>
#> 88   98      46 2015-11-04   999    <NA>     999
#> 89   97      46 2018-10-17   999     999    <NA>
#> 90   38      47 2018-09-21  3057    3044     999
#> 91    1      48 2016-10-29  3053    3044     999
#> 92   82      48 2018-04-22   999    <NA>    <NA>
#> 93   23      49 2016-02-22  3058    3042     999
#> 94   22      49 2018-06-18  3049    3045     999
#> 95   15      49 2020-03-28  3042    3050     999
#> 96   34      49 2020-04-08  3046    3043     999
#> 97   19      49 2020-09-01  3047     305    <NA>
#> 98   61      50 2015-08-06   999     999     999
#> 99   50      50 2016-03-11  3043    3052    3043
#> 100  86      50 2016-08-18   999     999     999