Grouping records into episodes by date range

This function is useful for collapsing, e.g., medication dispensation or hospitalization, records into episodes if the records' dates are no more than n days gap apart. The length of the gap can be relaxed by another grouping variable.

Usage

collapse_episode(
  data,
  clnt_id,
  start_dt,
  end_dt = NULL,
  gap,
  overwrite = NULL,
  gap_overwrite = Inf,
  .dt_trans = data.table::as.IDate,
  ...
)

Arguments

data

A data.frame that contains the id and date variables.

clnt_id

Column name of subject/person ID.

start_dt

Column name of the starting date of records.

end_dt

Column name of the end date of records. The default is NULL assuming the record last one day and only the start date will be used to calculate the gaps between records.

gap

A number in days that will be used to separate episodes. It should be the inclusive upper bound + 1. For example, gap = 8 means collapsing records no more than 7 days apart.

overwrite

Column name of a grouping variable determining whether the consecutive records are related and should have a different gap value. For example, dispensing records may have the same original prescription number, and a different gap value can be assigned for situation, e.g., the days between two records is > gap, but these records still belong to the same prescription.

gap_overwrite

A different gap value used for related records. The default is Inf, which means all records with the same overwrite variable will be collapsed.

.dt_trans

Function to transform start_dt/end_dt. Default is [data.table::as.IDate()].

...

Additional arguments passing to the .dt_trans function.

Value

The original data.frame with new columns indicating episode grouping

Examples

# make toy data
df <- make_test_dat()

head(df)
#>   uid clnt_id      dates diagx diagx_1 diagx_2
#> 1   9       3 2019-11-04  3046    <NA>    <NA>
#> 2  30       4 2017-01-15  3043    <NA>    <NA>
#> 3  47       5 2015-04-19  3049     304    3059
#> 4  61       5 2016-12-25   999     999    <NA>
#> 5  24       5 2017-12-20   305    3042     999
#> 6  63       5 2020-02-08   999    <NA>     999

# collapse records no more than 90 days apart
collapse_episode(df, clnt_id, dates, gap = 91)
#>     uid clnt_id      dates diagx diagx_1 diagx_2 epi_id epi_no epi_seq
#> 1     9       3 2019-11-04  3046    <NA>    <NA>      1      1       1
#> 2    30       4 2017-01-15  3043    <NA>    <NA>      2      1       1
#> 3    47       5 2015-04-19  3049     304    3059      3      1       1
#> 4    61       5 2016-12-25   999     999    <NA>      4      2       1
#> 5    24       5 2017-12-20   305    3042     999      5      3       1
#> 6    63       5 2020-02-08   999    <NA>     999      6      4       1
#> 7    82       6 2015-06-06   999     999    <NA>      7      1       1
#> 8    46       6 2017-02-16  3056    3041    3045      8      2       1
#> 9    43       6 2019-10-11   305    3053    3044      9      3       1
#> 10   44       7 2017-05-27  3046    3044    3048     10      1       1
#> 11   45       7 2018-04-16  3055    3058    3047     11      2       1
#> 12   95       7 2019-02-21   999     999     999     12      3       1
#> 13   81       7 2020-06-28   999     999     999     13      4       1
#> 14   57       8 2015-07-26   999    <NA>     999     14      1       1
#> 15   49       8 2018-03-10  3054    3044    3049     15      2       1
#> 16   90      10 2017-11-23   999    <NA>     999     16      1       1
#> 17   55      10 2018-08-31   999     999     999     17      2       1
#> 18   91      10 2019-02-09   999     999    <NA>     18      3       1
#> 19   79      11 2017-08-01   999    <NA>    <NA>     19      1       1
#> 20   56      11 2017-10-18   999     999    <NA>     19      1       2
#> 21   53      11 2019-05-11   999     999    <NA>     20      2       1
#> 22   50      11 2020-12-25  3047    3054    3057     21      3       1
#> 23   65      12 2015-09-17   999     999    <NA>     22      1       1
#> 24   89      12 2016-02-18   999    <NA>    <NA>     23      2       1
#> 25   25      12 2017-06-15  3040    3058     999     24      3       1
#> 26   48      12 2017-11-03  3056    3040    3054     25      4       1
#> 27   97      13 2019-05-22   999    <NA>    <NA>     26      1       1
#> 28    6      13 2020-10-06  3052    3054    <NA>     27      2       1
#> 29   38      14 2016-04-02  3059    3043     999     28      1       1
#> 30   62      15 2015-05-04   999    <NA>    <NA>     29      1       1
#> 31   52      16 2020-10-19   999    <NA>     999     30      1       1
#> 32   10      17 2015-01-17  3042    3048     999     31      1       1
#> 33   71      17 2015-04-06   999     999    <NA>     31      1       2
#> 34   60      17 2019-03-21   999     999    <NA>     32      2       1
#> 35   18      18 2015-02-10  3058     305    <NA>     33      1       1
#> 36   58      18 2017-03-06   999    <NA>    <NA>     34      2       1
#> 37   74      18 2018-10-07   999    <NA>    <NA>     35      3       1
#> 38   31      19 2018-05-15  3046    3043     999     36      1       1
#> 39   36      19 2018-06-07  3054     304    <NA>     36      1       2
#> 40   40      20 2015-07-22  3043    3058     999     37      1       1
#> 41   23      20 2016-11-29  3053    3043     999     38      2       1
#> 42   29      20 2020-08-23  3058    <NA>    <NA>     39      3       1
#> 43   83      21 2017-03-04   999     999     999     40      1       1
#> 44   32      21 2017-10-26  3059    3048    <NA>     41      2       1
#> 45   17      21 2018-10-06  3042    3046    <NA>     42      3       1
#> 46   34      21 2020-01-04  3058    3045    <NA>     43      4       1
#> 47   42      21 2020-02-24  3041    3050    3049     43      4       2
#> 48   33      22 2016-01-18  3047    3058    <NA>     44      1       1
#> 49   59      22 2020-10-17   999    <NA>    <NA>     45      2       1
#> 50   27      23 2018-12-13  3048     304    <NA>     46      1       1
#> 51   76      23 2019-04-04   999     999    <NA>     47      2       1
#> 52    5      23 2020-08-15  3052    3051     999     48      3       1
#> 53   78      23 2020-08-17   999    <NA>    <NA>     48      3       2
#> 54   21      24 2015-12-23  3043    3046    <NA>     49      1       1
#> 55   87      24 2018-05-01   999     999    <NA>     50      2       1
#> 56    2      25 2016-12-13  3049    3055    <NA>     51      1       1
#> 57   67      25 2018-06-15   999    <NA>    <NA>     52      2       1
#> 58   41      25 2019-01-24  3048    3049    3043     53      3       1
#> 59   80      26 2016-06-20   999     999    <NA>     54      1       1
#> 60  100      26 2018-01-06   999    <NA>     999     55      2       1
#> 61    8      26 2019-01-13  3043    3044    <NA>     56      3       1
#> 62   20      27 2020-08-05   305    3040     999     57      1       1
#> 63   92      27 2020-09-23   999     999     999     57      1       2
#> 64   39      28 2015-09-20   304    3048    <NA>     58      1       1
#> 65   64      28 2017-07-10   999    <NA>     999     59      2       1
#> 66   70      28 2019-06-19   999    <NA>     999     60      3       1
#> 67   98      28 2020-01-07   999     999     999     61      4       1
#> 68    1      29 2016-12-06  3050     304     999     62      1       1
#> 69   68      29 2020-02-25   999    <NA>    <NA>     63      2       1
#> 70    7      31 2017-01-07  3041    3041    <NA>     64      1       1
#> 71   77      32 2016-07-21   999     999    <NA>     65      1       1
#> 72   51      32 2018-11-19   999     999    <NA>     66      2       1
#> 73   72      33 2017-10-07   999    <NA>     999     67      1       1
#> 74   11      33 2020-01-31  3042    3055    <NA>     68      2       1
#> 75   93      34 2016-12-12   999    <NA>    <NA>     69      1       1
#> 76   22      34 2017-05-19  3041    <NA>    <NA>     70      2       1
#> 77    4      34 2017-06-26  3043    <NA>    <NA>     70      2       2
#> 78   54      34 2017-08-22   999     999    <NA>     70      2       3
#> 79   88      35 2019-09-09   999     999     999     71      1       1
#> 80   66      38 2020-09-10   999    <NA>    <NA>     72      1       1
#> 81   85      41 2016-07-11   999     999    <NA>     73      1       1
#> 82   73      41 2018-03-30   999    <NA>     999     74      2       1
#> 83   12      41 2019-08-19  3052    3052    <NA>     75      3       1
#> 84   19      42 2016-02-01  3049    3042    <NA>     76      1       1
#> 85   14      42 2017-01-23   305     305     999     77      2       1
#> 86   84      42 2017-11-06   999    <NA>     999     78      3       1
#> 87   15      42 2018-08-07  3048     304     999     79      4       1
#> 88   94      42 2020-11-17   999     999    <NA>     80      5       1
#> 89   16      43 2015-01-16  3048    3059    <NA>     81      1       1
#> 90   96      44 2015-10-17   999     999     999     82      1       1
#> 91   26      44 2018-11-28  3059    3055    <NA>     83      2       1
#> 92    3      45 2015-02-11  3042    3050    <NA>     84      1       1
#> 93   28      45 2015-03-23  3044    3044    <NA>     84      1       2
#> 94   37      46 2017-11-25  3057    3042    <NA>     85      1       1
#> 95   86      46 2019-12-08   999    <NA>    <NA>     86      2       1
#> 96   69      47 2019-08-26   999    <NA>     999     87      1       1
#> 97   35      48 2016-07-01   305    3046    <NA>     88      1       1
#> 98   13      48 2017-01-06  3050    3047    <NA>     89      2       1
#> 99   99      48 2017-07-29   999     999    <NA>     90      3       1
#> 100  75      48 2020-09-15   999    <NA>    <NA>     91      4       1
#>     epi_start_dt epi_stop_dt
#> 1     2019-11-04  2019-11-04
#> 2     2017-01-15  2017-01-15
#> 3     2015-04-19  2015-04-19
#> 4     2016-12-25  2016-12-25
#> 5     2017-12-20  2017-12-20
#> 6     2020-02-08  2020-02-08
#> 7     2015-06-06  2015-06-06
#> 8     2017-02-16  2017-02-16
#> 9     2019-10-11  2019-10-11
#> 10    2017-05-27  2017-05-27
#> 11    2018-04-16  2018-04-16
#> 12    2019-02-21  2019-02-21
#> 13    2020-06-28  2020-06-28
#> 14    2015-07-26  2015-07-26
#> 15    2018-03-10  2018-03-10
#> 16    2017-11-23  2017-11-23
#> 17    2018-08-31  2018-08-31
#> 18    2019-02-09  2019-02-09
#> 19    2017-08-01  2017-10-18
#> 20    2017-08-01  2017-10-18
#> 21    2019-05-11  2019-05-11
#> 22    2020-12-25  2020-12-25
#> 23    2015-09-17  2015-09-17
#> 24    2016-02-18  2016-02-18
#> 25    2017-06-15  2017-06-15
#> 26    2017-11-03  2017-11-03
#> 27    2019-05-22  2019-05-22
#> 28    2020-10-06  2020-10-06
#> 29    2016-04-02  2016-04-02
#> 30    2015-05-04  2015-05-04
#> 31    2020-10-19  2020-10-19
#> 32    2015-01-17  2015-04-06
#> 33    2015-01-17  2015-04-06
#> 34    2019-03-21  2019-03-21
#> 35    2015-02-10  2015-02-10
#> 36    2017-03-06  2017-03-06
#> 37    2018-10-07  2018-10-07
#> 38    2018-05-15  2018-06-07
#> 39    2018-05-15  2018-06-07
#> 40    2015-07-22  2015-07-22
#> 41    2016-11-29  2016-11-29
#> 42    2020-08-23  2020-08-23
#> 43    2017-03-04  2017-03-04
#> 44    2017-10-26  2017-10-26
#> 45    2018-10-06  2018-10-06
#> 46    2020-01-04  2020-02-24
#> 47    2020-01-04  2020-02-24
#> 48    2016-01-18  2016-01-18
#> 49    2020-10-17  2020-10-17
#> 50    2018-12-13  2018-12-13
#> 51    2019-04-04  2019-04-04
#> 52    2020-08-15  2020-08-17
#> 53    2020-08-15  2020-08-17
#> 54    2015-12-23  2015-12-23
#> 55    2018-05-01  2018-05-01
#> 56    2016-12-13  2016-12-13
#> 57    2018-06-15  2018-06-15
#> 58    2019-01-24  2019-01-24
#> 59    2016-06-20  2016-06-20
#> 60    2018-01-06  2018-01-06
#> 61    2019-01-13  2019-01-13
#> 62    2020-08-05  2020-09-23
#> 63    2020-08-05  2020-09-23
#> 64    2015-09-20  2015-09-20
#> 65    2017-07-10  2017-07-10
#> 66    2019-06-19  2019-06-19
#> 67    2020-01-07  2020-01-07
#> 68    2016-12-06  2016-12-06
#> 69    2020-02-25  2020-02-25
#> 70    2017-01-07  2017-01-07
#> 71    2016-07-21  2016-07-21
#> 72    2018-11-19  2018-11-19
#> 73    2017-10-07  2017-10-07
#> 74    2020-01-31  2020-01-31
#> 75    2016-12-12  2016-12-12
#> 76    2017-05-19  2017-08-22
#> 77    2017-05-19  2017-08-22
#> 78    2017-05-19  2017-08-22
#> 79    2019-09-09  2019-09-09
#> 80    2020-09-10  2020-09-10
#> 81    2016-07-11  2016-07-11
#> 82    2018-03-30  2018-03-30
#> 83    2019-08-19  2019-08-19
#> 84    2016-02-01  2016-02-01
#> 85    2017-01-23  2017-01-23
#> 86    2017-11-06  2017-11-06
#> 87    2018-08-07  2018-08-07
#> 88    2020-11-17  2020-11-17
#> 89    2015-01-16  2015-01-16
#> 90    2015-10-17  2015-10-17
#> 91    2018-11-28  2018-11-28
#> 92    2015-02-11  2015-03-23
#> 93    2015-02-11  2015-03-23
#> 94    2017-11-25  2017-11-25
#> 95    2019-12-08  2019-12-08
#> 96    2019-08-26  2019-08-26
#> 97    2016-07-01  2016-07-01
#> 98    2017-01-06  2017-01-06
#> 99    2017-07-29  2017-07-29
#> 100   2020-09-15  2020-09-15