Get variables from multiple tables with common ID columns

This function fetches variables from different tables that linked by common IDs. It calls dplyr::left_join() multiple times with different source tables (y argument of the join) to gather variables. It is not meant to replace left_join() but simplify syntax for the situation that you started off a table of study sample and wanted to gather covariates from different sources linked by common client IDs, which is often the case when working with healthcare databases. That said, this function is to replace repetitions of simple joins and only allows one-to-one matching.

Usage

fetch_var(data, keys, linkage, verbose = getOption("healthdb.verbose"), ...)

Arguments

data

A local data.frame, or tibble. It would be used as the x argument in left_join().

keys

A vector of quoted/unquoted variable names, or 'tidyselect' expression (see dplyr::select()). These variables must be present in data and would be used as the by argument in left_join(). The y tables must have a subset of these if not all.

linkage

A list of formulas in the form of "from_tab ~ get_vars|by_keys":

source table on the left-hand-side

variables on the right-hand-side

If a source table does not have all the variables in keys, use "|" on RHS to specify the subset of keys to be used.

For example, given keys has 3 variables, list( y1 ~ tidyselect_expr1, y2 ~ tidyselect_expr2|key1 + key2)

meaning:

from table y1 get variables picked by the tidyselect expression matching on all 3 keys;

from table y2 get variables matching on only key1 and key2.

verbose

A logical for whether report the number of rows after joining for each source. Default is getting from options. Use options(healthdb.verbose = FALSE) to suppress once and for all.

...

Additional arguments passing to left_join().

Value

A data.frame or tibble containing all original columns of x and new variables matched from other tables based on the specified linkage.

Examples

# make toy data
size <- 30
n <- 10
df1 <- data.frame(
  id = sample(1:n, size = size, replace = TRUE),
  service_dt = sample(seq(as.Date("2020-01-01"), as.Date("2022-01-31"), by = 1),
    size = size
  )
) %>%
  dplyr::mutate(year = lubridate::year(service_dt))
df2 <- data.frame(
  id = rep(1:n, size / n), year = rep(2020:2022, each = n),
  status_1 = sample(0:1, size = size, replace = TRUE),
  status_2 = sample(0:1, size = size, replace = TRUE)
)
df3 <- data.frame(id = 1:n, sex = sample(c("F", "M"), size = n, replace = TRUE))

# simple joins
# note that for left_join(df1, df2), boths keys have to be used,
# otherwise, error as the relation would not be one-to-one
fetch_var(df1,
  keys = c(id, year),
  linkage = list(
    df2 ~ starts_with("s"), # match both keys without '|'
    df3 ~ sex | id
  ) # match by id only; otherwise failed because df3 has no year
)
#> 
#> The data has 30 rows. After joining, variable(s) from df2 has 30 rows, and variable(s) from df3 has 30 rows 
#>    id service_dt year status_1 status_2 sex
#> 1   3 2020-04-24 2020        0        1   F
#> 2   6 2021-12-12 2021        1        0   F
#> 3   1 2021-03-02 2021        0        1   F
#> 4  10 2020-06-12 2020        0        0   F
#> 5  10 2020-02-29 2020        0        0   F
#> 6   4 2020-07-21 2020        0        0   F
#> 7   9 2020-06-09 2020        0        1   M
#> 8   2 2021-07-29 2021        1        0   F
#> 9   5 2020-05-19 2020        0        1   F
#> 10  3 2021-08-01 2021        0        1   F
#> 11  6 2020-01-25 2020        1        1   F
#> 12  7 2021-03-23 2021        1        0   M
#> 13  2 2021-04-16 2021        1        0   F
#> 14  1 2021-04-21 2021        0        1   F
#> 15  2 2020-03-13 2020        0        0   F
#> 16 10 2020-04-06 2020        0        0   F
#> 17  7 2020-12-29 2020        0        1   M
#> 18  3 2022-01-24 2022        1        1   F
#> 19  6 2021-01-13 2021        1        0   F
#> 20  5 2021-04-23 2021        0        0   F
#> 21  9 2021-03-16 2021        1        1   M
#> 22  8 2020-05-09 2020        0        0   M
#> 23  2 2021-11-03 2021        1        0   F
#> 24  7 2021-04-27 2021        1        0   M
#> 25  4 2020-01-18 2020        0        0   F
#> 26  7 2021-12-21 2021        1        0   M
#> 27  5 2020-12-31 2020        0        1   F
#> 28  4 2020-03-24 2020        0        0   F
#> 29 10 2021-02-13 2021        1        0   F
#> 30  4 2021-02-16 2021        1        0   F

# example if some y is remote
# make df2 as database table
db2 <- dbplyr::tbl_memdb(df2)

fetch_var(df1,
  keys = c(id, year),
  linkage = list(
    db2 ~ starts_with("s"),
    df3 ~ sex | id
  ),
  copy = TRUE # pass to left_join for forced collection of remote table
)
#> 
#> The data has 30 rows. After joining, variable(s) from db2 has 30 rows, and variable(s) from df3 has 30 rows 
#>    id service_dt year status_1 status_2 sex
#> 1   3 2020-04-24 2020        0        1   F
#> 2   6 2021-12-12 2021        1        0   F
#> 3   1 2021-03-02 2021        0        1   F
#> 4  10 2020-06-12 2020        0        0   F
#> 5  10 2020-02-29 2020        0        0   F
#> 6   4 2020-07-21 2020        0        0   F
#> 7   9 2020-06-09 2020        0        1   M
#> 8   2 2021-07-29 2021        1        0   F
#> 9   5 2020-05-19 2020        0        1   F
#> 10  3 2021-08-01 2021        0        1   F
#> 11  6 2020-01-25 2020        1        1   F
#> 12  7 2021-03-23 2021        1        0   M
#> 13  2 2021-04-16 2021        1        0   F
#> 14  1 2021-04-21 2021        0        1   F
#> 15  2 2020-03-13 2020        0        0   F
#> 16 10 2020-04-06 2020        0        0   F
#> 17  7 2020-12-29 2020        0        1   M
#> 18  3 2022-01-24 2022        1        1   F
#> 19  6 2021-01-13 2021        1        0   F
#> 20  5 2021-04-23 2021        0        0   F
#> 21  9 2021-03-16 2021        1        1   M
#> 22  8 2020-05-09 2020        0        0   M
#> 23  2 2021-11-03 2021        1        0   F
#> 24  7 2021-04-27 2021        1        0   M
#> 25  4 2020-01-18 2020        0        0   F
#> 26  7 2021-12-21 2021        1        0   M
#> 27  5 2020-12-31 2020        0        1   F
#> 28  4 2020-03-24 2020        0        0   F
#> 29 10 2021-02-13 2021        1        0   F
#> 30  4 2021-02-16 2021        1        0   F