healthdb: Working with Healthcare Databases

A system for identifying diseases or events from healthcare databases and preparing data for epidemiological studies. It includes capabilities not supported by 'SQL', such as matching strings by 'stringr' style regular expressions, and can compute comorbidity scores (Quan et al. (2005) doi:10.1097/01.mlr.0000182534.19832.83 ) directly on a database server. The implementation is based on 'dbplyr' with full 'tidyverse' compatibility.

See also

Author

Maintainer: Kevin Hu kevin.hu@bccdc.ca (ORCID) [copyright holder]