Make test data

Make a toy data set for testing and demo. This is for internal use purpose and not intended to be called by users.

Usage

make_test_dat(
  vals_kept = c("304", "305", 3040:3049, 3050:3059),
  noise_val = "999",
  IDs = 1:50,
  date_range = seq(as.Date("2015-01-01"), as.Date("2020-12-31"), by = 1),
  nrows = 100,
  n_any = 50,
  n_all = 10,
  seed = 2024,
  answer_id = NULL,
  type = c("data.frame", "database")
)

Arguments

vals_kept

A vector of values that suppose to be identified.

noise_val

A vector of values that are not meant to be identified.

IDs

A vector of client IDs.

date_range

A vector of all possible dates in the data.

nrows

Number of rows of the output.

n_any

Number of rows to be identified if the criteria is that if any target column contains certain values.

n_all

Number of rows to be identified if the criteria is that if all target columns contain certain values.

seed

Seed for random number generation.

answer_id

Column name for the indicator of how the row should be identified: any, all, and noise.

type

Output type, "data.frame" or "database".

Value

A data.frame or remote table from 'dbplyr'.

Examples

make_test_dat()
#>     uid clnt_id      dates diagx diagx_1 diagx_2
#> 1    68       2 2016-08-03   999     999     999
#> 2    11       2 2020-07-02   305    3055    <NA>
#> 3    23       3 2019-07-08  3057    3056     999
#> 4    49       4 2017-10-23  3042    3058    3045
#> 5    89       4 2017-11-04   999     999     999
#> 6     9       5 2018-12-23  3053     304    <NA>
#> 7    18       6 2017-03-07  3046    3047    <NA>
#> 8    66       7 2019-07-17   999     999     999
#> 9    58       8 2017-06-08   999    <NA>    <NA>
#> 10   45       8 2018-02-22   304    3041    3056
#> 11   61       8 2020-10-26   999     999     999
#> 12   79       9 2019-05-03   999     999    <NA>
#> 13   85       9 2019-09-04   999    <NA>    <NA>
#> 14   95       9 2020-12-25   999    <NA>    <NA>
#> 15   15      10 2017-05-29   304    3042     999
#> 16   93      11 2015-08-25   999     999     999
#> 17    2      12 2016-01-25  3042    3056     999
#> 18   77      12 2018-04-22   999     999    <NA>
#> 19   54      12 2020-06-10   999     999     999
#> 20    3      13 2018-09-12  3050    3045    <NA>
#> 21   16      14 2015-12-29  3048     305    <NA>
#> 22   97      14 2020-04-30   999     999    <NA>
#> 23   70      14 2020-08-16   999     999     999
#> 24   17      15 2015-10-14  3040    <NA>    <NA>
#> 25   13      15 2017-07-23  3055    3041    <NA>
#> 26   12      15 2018-06-17  3049    3048     999
#> 27   83      16 2018-03-03   999    <NA>     999
#> 28   32      16 2020-05-17  3053    3052     999
#> 29   25      17 2017-01-29  3057    3052    <NA>
#> 30   44      17 2019-12-04  3059    3058     304
#> 31   22      18 2015-05-16  3057    3047    <NA>
#> 32   29      18 2017-03-06  3044    3044    <NA>
#> 33   20      18 2018-03-16  3054    3045     999
#> 34   43      18 2019-07-29  3058    3054    3047
#> 35   88      18 2019-12-29   999    <NA>     999
#> 36   10      19 2018-04-08  3051    3053     999
#> 37   59      19 2020-12-04   999     999    <NA>
#> 38   51      20 2016-01-20   999     999    <NA>
#> 39   31      20 2020-06-09  3058    3056    <NA>
#> 40   60      21 2015-01-19   999     999     999
#> 41   47      21 2016-12-05  3040    3046    3055
#> 42   41      21 2020-06-21   305    3058     305
#> 43   91      21 2020-11-16   999     999    <NA>
#> 44    5      22 2017-10-20  3050    3049     999
#> 45   71      23 2019-03-08   999     999    <NA>
#> 46   72      24 2019-10-10   999    <NA>     999
#> 47   35      24 2020-07-31   305    3047    <NA>
#> 48   84      28 2017-12-30   999    <NA>     999
#> 49   38      28 2018-01-15  3054    3048    <NA>
#> 50  100      29 2018-01-01   999    <NA>     999
#> 51   24      29 2019-11-24  3048    3050     999
#> 52    1      29 2020-12-17   305    3042    <NA>
#> 53   78      30 2015-11-19   999    <NA>     999
#> 54   50      30 2016-04-25  3042    3055    3045
#> 55   96      30 2017-11-14   999    <NA>    <NA>
#> 56    7      30 2019-11-05   305    3047    <NA>
#> 57   63      31 2016-03-24   999    <NA>    <NA>
#> 58   69      31 2016-04-09   999     999    <NA>
#> 59   52      31 2016-09-23   999    <NA>    <NA>
#> 60   75      31 2017-11-26   999    <NA>     999
#> 61   36      32 2020-10-14  3054     304     999
#> 62   94      33 2016-03-01   999    <NA>     999
#> 63   28      33 2016-07-24  3041    3053    <NA>
#> 64    8      33 2019-12-14  3048    3057    <NA>
#> 65   33      35 2017-09-26  3057    3058     999
#> 66   62      35 2018-08-14   999    <NA>    <NA>
#> 67   99      36 2015-12-02   999     999    <NA>
#> 68   80      36 2016-04-22   999     999    <NA>
#> 69   14      36 2018-08-13  3058    3052    <NA>
#> 70   55      36 2019-12-10   999     999     999
#> 71   82      38 2016-11-29   999     999    <NA>
#> 72   57      38 2017-01-25   999     999    <NA>
#> 73   46      39 2017-06-24   305    3053     304
#> 74   81      39 2017-09-23   999     999     999
#> 75    6      39 2018-11-27  3046    3052     999
#> 76   74      39 2020-02-05   999    <NA>    <NA>
#> 77   87      40 2020-07-14   999    <NA>    <NA>
#> 78   19      41 2018-01-14  3051    3055    <NA>
#> 79   64      42 2016-05-21   999     999     999
#> 80   65      42 2018-06-27   999     999     999
#> 81    4      43 2016-01-24  3054    3051     999
#> 82   37      43 2016-07-03  3059    3055     999
#> 83   40      43 2018-11-13   304     304     999
#> 84   21      44 2016-10-18  3053     305     999
#> 85   67      44 2017-04-21   999     999     999
#> 86   56      44 2019-07-17   999    <NA>     999
#> 87   53      45 2015-10-02   999    <NA>    <NA>
#> 88   26      45 2018-08-08  3057    3040    <NA>
#> 89   34      45 2019-07-31  3040    3054     999
#> 90   48      45 2019-08-20  3050    3059     304
#> 91   27      46 2020-02-23  3047    3050     999
#> 92   76      47 2015-11-05   999     999     999
#> 93   30      47 2018-02-21  3057    3045    <NA>
#> 94   42      48 2017-03-21  3048     305    3050
#> 95   73      48 2019-03-09   999    <NA>    <NA>
#> 96   39      48 2019-08-28  3040    3053    <NA>
#> 97   90      49 2015-08-26   999    <NA>    <NA>
#> 98   98      49 2015-09-03   999     999    <NA>
#> 99   86      49 2020-08-17   999     999     999
#> 100  92      50 2019-07-24   999    <NA>     999